move populateDarData and populateProgressReport to DataAccessRequestData and DataAccessRequest respectively
